test: 테스트에 H2반영
