+++ backend/src/test/java/codesquad/team4/issuetracker/issue/IssueServiceTest.java
@@ -2,6 +2,7 @@ package codesquad.team4.issuetracker.issue;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
